Bugs!! Fixes! Spring cleaning!
Because these are either newly created, or newly discovered, bugs,
this release was pushed as early as possible to get them fixed asap.

BUGS:
1. Desktop: Lumina detection had a syntax error which made it not work.
This has been broken for a while.

2. Logical: if not root, and if LUKS / bcache detected, failed to load
proc_partitions, which generates error on --logical --admin since the
required components data was not loaded. This was an oversight.

3. The 3.3.02 ShellData refactor created a bug for console IRC, showed
shell, not irc client, name, and set default shell data which also
showed.

4. Console IRC tty: there was also an older bug that made -S, -G
not work consistently, and there were errors that had been missed
for many years in that logic. These should all be corrected, console
irc out of display, or in display as root, should now show tty info,
tty size in -G.

FIXES:
1. Memory: restored $bsd_type block on /proc/meminfo and force NetBSD
to use a corrected vmstat. This leaves that block of logic to correct
the NetBSD oddities in meminfo, but it may fix future isses that popup.

2. -Sxxx man page item incorrectly said XDG_VTNR was systemd/linux, it's
not, it's various things, GhostBSD has it, for example. See what you get
for believing what people say!

3. Logical: added in N/A for null maj-min in --logical report. While
bug 2 triggered those errors, there could be future cases where maj-min
are null, like BSD lvm data etc.

ENHANCEMENTS:
1. Going along with Fix 1, added '--force meminfo' in case you really want that.

2. Distro: System Base: Added TrueNAS detection.

3. Package Data: Added mport [MidnightBSD] type. That requires root to run
for some odd reason, so won't see the best output if not root.

CHANGES:
1. Moved logical to -v7 from -v8, it's stable enough now.

CODE CHANGES:
1. Moved get_tty_number and get_tty_console_irc to ShellData:tty_number
and ShellData::tty_console_irc.

ShellData::tty_number was being loaded several times, added
$loaded{'tty-number'} test, and made client{'tty-number'} to store value.

tty_console_irc changed to console_irc_tty, which is what it gets, removed
hacks and made it load once and store result in client hash.

2. Optimization: retested sub vs package::method and they run at exactly
the same time, give or take, so moving more stuff into packages to make
it easier to maintain.