Bug fix release. 2 bugs that can impact all users under the right circumstances were detected and fixed. Thanks manjaro users there for finding and reporting those. No other changes.
--------------------------------------------------------------------------------
BUGS:

1. Manjaro user ben81 located a critical bug in hardware raid output, this bug
impacts ALL users of hwraid that run inxi with -xx option. Bug was a bad copy
paste, the classic, had updated all the pci type data blocks at once, and hw
raid unfortunately had a slightly different logic due to being part of the more
complex RAID block of logic. Was trying to use an array, not a hash, reference.

Thanks ben81, I would never have spotted this one, and it would impact 100% of
all inxi users with hwraid on their machine who ran inxi with -xx option.

2. Also, ps wwaux parser was spitting out an undefined index error. This is
caused by one of two things:

* ps has an issue, and is apparently at times failing to respect ww, unlimited
line length, and wrapping anyway. This is the likely cause.
* the user terminal for some inexpicable reason has decided to hard wrap long
lines. This is very unlikely, but has to be considered as a possible cause.
Since these commands run in a subshell, this is VERY unlikely.

Workaround this failure by double checking that line split item is defined, if
not, next row. Thanks Carpenter for finding that one.

--------------------------------------------------------------------------------
CODE:

1. Added workarounds for bug 2. Corrected silly copy/paste error for bug 1.