Quick bug fix release. With as many changes as we got in 3.3.10, there were bound to be a handful of oversights that were not caught in testing simply because those hardware scenarios were not present in the tested systems.
Also minor feature enhancement for CPU scaling min/max speeds.

--------------------------------------------------------------------------------
KNOWN ISSUES:

1. Due to the huge amount of changes, and the speed of change, while the new
code is working as intended, it's somewhat lacking in coding elegance since a
lot of it was hacked out very quickly, in near real time. This will get cleaned
up and refactored to be less redundant if it does not impact execution speed,
but is not pressing since there should be no functional difference.

--------------------------------------------------------------------------------
BUGS:

1. Tiny oversight, in single case CPU model id would fail because it was using
an undefined test from previous tests, not the right test, that is. Tripped
error on Elbrus for example.

2. Typo in battery secondary type status, created undefined value error. This
was I believe an older bug.

--------------------------------------------------------------------------------
FIXES:

1. PPC revision change broke Elbrus revision test for stepping. Added in more
tests to show stepping for elbrus revision.

2. Single core Elbrus in cpuinfo fallback mode failed to assign core multiplier
so L1 cache failed.

3. In cpuinfo fallback mode, Elbrus E2C3 cache data failed to appear, that data
in not per block in cpuinfo, but is the last block, so those tests had to run
on each block, not just the first one.

--------------------------------------------------------------------------------
ENHANCEMENTS:

1. Show for -Ca scaling min/max speeds if different from CPU min/max speeds.

2. If no cpuinfo_min/max_freq speeds found, and scaling_min/max_freq found, set
overall min/max to use scaling min/max instead of cpuinfo min/max.

--------------------------------------------------------------------------------
CHANGES:

1. None.

--------------------------------------------------------------------------------
DOCUMENTATION:

1. Cleaned up and proofread better 3.3.10 changelog, it had a lot of errors
because stuff kept changing.

--------------------------------------------------------------------------------
CODE:

1. Small code optimizations.