Small point release with some useful fixes, some bugs corrected, small enhancement.
--------------------------------------------------------------------------------
KNOWN ISSUES:

1. Note on KNOWN ISSUES 2 from 3.3.10, I realized that the second level
indentation actually takes care of that hanging parent, since now it's quite
clear that its children are indented on the following line. So that issue took
care of itself.

--------------------------------------------------------------------------------
BUGS:

1. In $bsd_type cases, specifically SunOS triggered this, where sysctl was not
present (which it always is in all other BSD types), it tripped an error due to
failure to update to revised $alerts{'sysctl'}->{'message'}, instead was using
the old method. Failed to update that in refactor of CheckTools logic. Would
impact no supported operating systems, but is a bug.

2. Corner case, a combination I never used, inxi -a, triggers error in the RAID
logic because the mdadm test was not run, generating an undefined eq error. Only
happened when mdraid was present and used on system.

3. Tiny bug, for st+mt cpus, like alder lake,  was printing out second tpc by
accident which made it look like it said st: 4 tpc: 2. Just a small output
glitch.

--------------------------------------------------------------------------------
FIXES:

1. Added some fallback 32 bit system tests, [2345]86, like i386, i686.

2. Changed shell: Unknown Shell to shell: N/A, that was a legacy use, and was
obviously redundant. N/A is more consistent with rest of output no value found
handling. Due to requirement of doing empty tests, this is set in the data, not
output, generator component.

3. Refactored partition data logic to get rid of bsd tests for df -kTP, -kT, -k.
Now tests only to see if returns data, cascades down until it gets something.
Now will attempt to reconnect hanging lines when no -P feature available prior
to main partition data processing. This makes it agnostic to os issues, and it
just pays attention to feature support.

Also adds in dynamic column count instead of hard-coded, this avoids oddities
and future proofs to some degree. Now systems will adapt seamlessly if support
for -P appears, or -T, or whatever.

4. Partitions corner case, where has zram, but has no partitions, failed to show
partitions no data found message since @partitions had data in it, but nothing
for partitions output to print. Seen in TinyCore for example, but might happen
in other ram based systems.

--------------------------------------------------------------------------------
ENHANCEMENTS:

1. Added Slint to distro ID, and slint/slackware to system base.

--------------------------------------------------------------------------------
CHANGES:

1. None

--------------------------------------------------------------------------------
DOCUMENTATION:

1. Typo in man page fixed.

2. A few more edits and corrections on 3.3.10 changelong. That thing was written
concurrently with the development, and thus had lingering errors when things
were changed in midstream.

--------------------------------------------------------------------------------
CODE:

1. Added --cygwin and --android fake system type flag. Switches on $b_android/
$b_cygwin flags.