A small point release, some nice enhancements, a few bug fixes, and some fixes to correct or update some issues. Main new features: * -rx,-rxx,-ra/-Ix,-Ixx,-Ia: enhanced Packages: report * -Cxx: microarchitecture level. This is a relatively new convention, used to set various compile flags.
--------------------------------------------------------------------------------
KNOWN ISSUES:

1. Can't safely look for pm tool KDE Discover because apt calls it
plasma-discover, but other packaging systems call it discover, which is already
a non-related program (hardware data). Since it's not really core to any package
manager, it's not really a necessary thing to report anyway, though
gnome-software is added because that appears to be more like syntaptic than
anything else.

There's also a qt variant of the rpm packagekit, packagekit-qt, which is
available in for example Arch, but again, it's too granular, and not really
core.

2. At some point, sensors should add /sys hwmon sensor data, then switch to
using lm-sensors as a fallback, and remove one recommended tool from newer linux
systems. I don't think that's too hard, just a bunch of little steps to
integrate that into the main logic.

--------------------------------------------------------------------------------
BUGS:

1. For Slackware slackpkg/pkgtool:
a: Failed to show package counts at all because of bad globbing path, forgot /*.
b. Failed to show lib counts for packages due to having wrong counter for path.

2. If no ipmi sensor data was found but the tools are present, could result in
an undefined hash reference error for sensors. The most likely cause for this is
that one of the ipmi commands:
"ipmi-tool sensors" or "ipmi-sensors"
had an error, and since errors are sent to /dev/null, inxi saw null data, then
returned an undefined value instead of the hash reference it was supposed to.
This is the first time I've seen this happen with ipmi, but there have not been
a lot of ipmi samples.

Thanks issue poster #274 for having systems that triggered this scenario.

3. $source for ipmi was set to lm-sensors by accident.

4. For sensors, with > 1 sensor type, like lm-sensors + ipmi, sensor data from
second sensor type was getting written to first sensor type row. See Fixes 9 for
more sensors fixes.

--------------------------------------------------------------------------------
FIXES:

1. Force CPU bits to 64 if LM flag is present, if it reports as i686. This fix
only runs for non RISC CPUs that show as 32 bit, so it won't run very often. If
no LM found, remains 32 bit.

This fix goes along with enhancement 1, which only applies to 64 bit CPUs.

2. In --recommends, JSON::PP module package names were wrong, they were copied
from JSON::Cpanel::XS and hadn't been changed to the right package names. Note
that for most distros, this is in Perl Core Modules, but not all.

3. Samsung ram vendor id was too tight, loosened it up a bit. Missed this one:
K3LK7K70BM

4. With Bug 1, extended possible package manager tool detection for slackware
type systems. Slackware is kind of unique in that it is not actually made out of
a core package manager as a collection of packages, but uses package managers as
a kind of layer on top of that, but none of those tools is required to run the
system.

5. Found another corner case indentation glitch, was adding in level 2 on -I
which is has no second level indentation.

6. Forgot to add $force{'pkg'} to -v8.

7. Small fix, if -Z is used, forgot to force --zl, --zu, --zv to false as well
as -z.

8. Small fix, for saphire rapids, alder lake, added + to year built, since those
are ongoing.

9. Sensors:
a. in one case, with an array of fan speeds, set to '' instead of undef, which
made test fail, and showed empty fan item.
b. added wildcards for possible voltage/power matches, was too restrictive for
ipmi sensors values.
c. added better space regex for ipmi temps [\s_-]?
d. DIMM voltage/temp excluded > 9 numbering, like DIMM 19

--------------------------------------------------------------------------------
ENHANCEMENTS:

1. New feature: -Cxx shows for AMD / Intel 64 bit CPUs the microarchitecture
level (v1,v2,v3,v4). v1 is baseline. GCC supports this I believe in latest
versions, and some distros use it to determine CPU support levels for compile
time optimizations. This was introduced in 2020 via a collaboration between AMD,
Intel, SUSE, and Redhat. Now you know. This is a simple test based on which CPU
flags/features are present.

These levels can be used for Go language optimization (GOAMD64), GCC
optimization switches (GCC -O2 for example), and probably more.

2. Expanded YMTC (Yangtze Memory) RAM vendor ids and detections.

3. Added [unverified] window managers CDE and NsCDE. No data, only using ps aux
method.

4a. Added slax ID to distro id, added slax to system base support. Currently
only work on slackware based 15.0, not debian based 11.4.

4b. Added SteamOS debian/arch for system  base.

4c. Added os-release VERSION_CODENAME to enhance distro ID data (eg steamos)

5. Added to -ra/-Ia package tools installed report, this goes along with change
2, which changes apt to dpkg, the low level tool. Now with -a, shows the package
manager tools installed, like slackpkg, apt, apt-get, dnf, yum, zypper, etc.

rpm installed as secondary pm requires some further tests. Currently known pm
that have tools (and rpm tests if detected):
All these are known to support rpm secondary pm:
* dpkg - Debian, Ubuntu, and apt-rpm based distros like PCLinuxOS, Alt Linux
* pacman - Arch based distros
* pkgtool - Slackware based distros
* tce - TinyCore Linux

6. A few more pci product IDs for GPU matches. Slow going.

--------------------------------------------------------------------------------
CHANGES:

1. Changed --pkg to --rpm, the original intention was that this could apply to
more than RPM package manager, but that's the only one that it's used for. This
leads to unclear output for other distributions where the user might have rpm
installed alongside their standard package manager.

2. Changed package pm: 'apt' to 'dpkg', to go along with type rpm (suse,redhat)
and pkgtool (slackware). Note that dpkg is the actual package manager of Debian,
inxi had this wrong, apt interacts with dpkg.

3. Changed -h -a section, to follow after -x, -xx, -xxx, like on man page.

4. For rpm notes, after running some tests to determine whether to use rpm or
not, will show the rpm note: see --rpm in pm: rpm note:... This allows for
more granular errors which will be more useful to users.

--------------------------------------------------------------------------------
DOCUMENTATION:

1. Docs were wrong for -ra/-Ia packages, from original when it package report
was only an -a option, but it got moved to -rx, -rxx for basic features, and -a
for advanced features.

2. Updated for --pkg/--rpm and --force rpm/pkg

3. See change 3, I think people tend to miss the sequence of -x, -xx, -xxx, -a
because -a came before -x, -xx, -xxx in -h menu, but on man page, -a correctly
comes after the -xxx options. Better to be consistent.

--------------------------------------------------------------------------------
CODE:

1. Switched force{pkg} to force{package} internally, and added converts to
change --force pkg/rpm to switch on $force{package}.

2. Refactored package PackageData to be more granular.